+++ app/build.gradle
@@ -69,7 +69,7 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     runtimeOnly 'com.h2database:h2'
 
-//    implementation "org.springframework.boot:spring-boot-starter-security"
+    implementation "org.springframework.boot:spring-boot-starter-security"
 
     // Spring Developer Tools
     developmentOnly 'org.springframework.boot:spring-boot-devtools'
