--- app/src/main/java/com/codesoom/assignment/config/WebJavaConfig.java
@@ -1,18 +1,0 @@
-package com.codesoom.assignment.config;
-
-import com.codesoom.assignment.interceptors.AuthenticationInterceptor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebJavaConfig implements WebMvcConfigurer {
-    @Autowired
-    private AuthenticationInterceptor authenticationInterceptor;
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(authenticationInterceptor);
-    }
-}
