+++ app/src/main/java/com/codesoom/assignment/controllers/ControllerErrorAdvice.java
@@ -39,11 +39,11 @@ public class ControllerErrorAdvice {
         return new ErrorResponse("Log-in failed");
     }
 
-    @ResponseStatus(HttpStatus.UNAUTHORIZED)
-    @ExceptionHandler(InvalidTokenException.class)
-    public ErrorResponse handleInvalidAccessTokenException() {
-        return new ErrorResponse("Invalid access token");
-    }
+//    @ResponseStatus(HttpStatus.UNAUTHORIZED)
+//    @ExceptionHandler(InvalidTokenException.class)
+//    public ErrorResponse handleInvalidAccessTokenException() {
+//        return new ErrorResponse("Invalid access token");
+//    }
 
     @ResponseBody
     @ResponseStatus(HttpStatus.BAD_REQUEST)
